backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Database ──────────────────────────────────────────────
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# ── Auth ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# ── AI / RAG ──────────────────────────────────────────────
anthropic==0.34.2
voyageai==0.2.3
pinecone-client==4.1.2

# ── Document Parsing ──────────────────────────────────────
pypdf==4.3.1
python-docx==1.1.2

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2